+++ lib/data/game/entities/game.dart
@@ -1,0 +1,26 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'game.freezed.dart';
+part 'game.g.dart';
+
+@Freezed(makeCollectionsUnmodifiable: false)
+class Games with _$Games {
+  const factory Games({
+    required int id,
+    required String name,
+    required int platformId,
+  }) = _Games;
+
+  factory Games.fromJson(Map<String, dynamic> json) => _$GamesFromJson(json);
+}
+
+@Freezed(makeCollectionsUnmodifiable: false)
+class GamesList with _$GamesList {
+  const factory GamesList({
+    required List<Games> games,
+  }) = _GamesList;
+  const GamesList._();
+
+  factory GamesList.fromJson(Map<String, dynamic> json) =>
+      _$GamesListFromJson(json);
+}