+++ lib/data/queries.dart
@@ -1,0 +1,19 @@
+String query = '''
+    query {
+  platforms{
+   
+    nodes{
+      name
+      id
+      games{
+        nodes{
+          name
+          id
+          platformId 
+        }
+      }
+      
+    }
+  }
+}
+''';